Add JS Async implementation and make it the default
This package only provides a synchronous version of each function and especially for the `slowAfter` this might be a problem because it uses a while loop for sleeping.

This Commit renames the existing functions to their *Sync version and adds async implementations (which most of the time just use the sync one).

It also makes the async ones the default, so a new developer would be encouraged to use those functions instead.

Tests are added to reflect the sync versions and also readme is updated.

Signed-off-by: Raphael Höser <raphael@hoeser.info>